Fix int overflow in array index arithmetic for large matrices
Multiple places computed `int * int` (e.g. `j * ldim_A`) without a
`(size_t)` cast. For matrices larger than ~46340x46340, the product
exceeds INT_MAX, wrapping to a negative value and producing an
out-of-bounds pointer. Writing through such a pointer corrupts heap
metadata, causing subsequent free() calls to crash or misbehave
(double-free / invalid-free symptoms).

Cast the first operand to `(size_t)` in all affected index expressions
across dgeqp4, NoFLA_HQRRP_WY_blk_var4, NoFLA_QRPmod_WY_unb_var4,
NoFLA_QRP_pivot_G_B_C, and matrix_generate in simple_test.c.

+++ lapack_compatible_sources/simple_test.c
@@ -130,7 +130,7 @@ static void matrix_generate( int m_A, int n_A, double * buff_A, int ldim_A ) {
 
   for ( j = 0; j < n_A; j++ ) {
     for ( i = 0; i < m_A; i++ ) {
-      buff_A[ i + j * ldim_A ] = 2.0 * ( ( double ) rand() / RAND_MAX ) - 1.0;
+      buff_A[ i + (size_t)j * ldim_A ] = 2.0 * ( ( double ) rand() / RAND_MAX ) - 1.0;
     }
   }
 }
